All Hydra fixes and improvements.

+++ Cargo.hooks.php
@@ -21,15 +21,6 @@ class CargoHooks {
 			'Datetime', 'Boolean', 'Coordinates', 'Wikitext',
 			'Searchtext', 'File', 'URL', 'Email'
 		);
-
-		$wgGroupPermissions['sysop']['recreatecargodata'] = true;
-		$wgGroupPermissions['sysop']['deletecargodata'] = true;
-		$wgGroupPermissions['*']['runcargoqueries'] = true;
-
-		// Backward compatibility for MW < 1.28.
-		if ( !defined( 'DB_REPLICA' ) ) {
-			define( 'DB_REPLICA', DB_SLAVE );
-		}
 	}
 
 	public static function registerParserFunctions( &$parser ) {
@@ -165,7 +156,7 @@ class CargoHooks {
 		// Get all the "main" tables that this page is contained in.
 		$dbw = wfGetDB( DB_MASTER );
 		$cdb = CargoUtils::getDB();
-		$cdb->begin();
+		$cdb->startAtomic( __METHOD__ );
 		$cdbPageIDCheck = array( $cdb->addIdentifierQuotes( '_pageID' ) => $pageID );
 
 		$res = $dbw->select( 'cargo_pages', 'table_name', array( 'page_id' => $pageID ) );
@@ -181,16 +172,18 @@ class CargoHooks {
 			$res2 = $dbw->select( 'cargo_tables', 'field_tables', array( 'main_table' => $curMainTable ) );
 			$row2 = $dbw->fetchRow( $res2 );
 			$fieldTableNames = unserialize( $row2['field_tables'] );
-			foreach ( $fieldTableNames as $curFieldTable ) {
-				// Thankfully, the MW DB API already provides a
-				// nice method for deleting based on a join.
-				$cdb->deleteJoin(
-					$curFieldTable,
-					$curMainTable,
-					$cdb->addIdentifierQuotes( '_rowID' ),
-					$cdb->addIdentifierQuotes( '_ID' ),
-					$cdbPageIDCheck
-				);
+			if ( is_array( $fieldTableNames ) ) {
+				foreach ( $fieldTableNames as $curFieldTable ) {
+					// Thankfully, the MW DB API already provides a
+					// nice method for deleting based on a join.
+					$cdb->deleteJoin(
+						$curFieldTable,
+						$curMainTable,
+						$cdb->addIdentifierQuotes( '_rowID' ),
+						$cdb->addIdentifierQuotes( '_ID' ),
+						$cdbPageIDCheck
+					);
+				}
 			}
 
 			// Delete from the "files" helper table, if it exists.
@@ -211,7 +204,7 @@ class CargoHooks {
 		$dbw->delete( 'cargo_pages', array( 'page_id' => $pageID ) );
 
 		// End transaction and apply DB changes.
-		$cdb->commit();
+		$cdb->endAtomic( __METHOD__ );
 	}
 
 	/**
@@ -311,7 +304,7 @@ class CargoHooks {
 		$newPageNamespace = $newtitle->getNamespace();
 		$dbw = wfGetDB( DB_MASTER );
 		$cdb = CargoUtils::getDB();
-		$cdb->begin();
+		$cdb->startAtomic( __METHOD__ );
 		// We use $oldid, because that's the page ID - $newid is the
 		// ID of the redirect page.
 		// @TODO - do anything with the redirect?
@@ -344,7 +337,7 @@ class CargoHooks {
 		}
 
 		// End transaction and apply DB changes.
-		$cdb->commit();
+		$cdb->endAtomic( __METHOD__ );
 
 		return true;
 	}
